Partially fix circular dependencies
