+++ src/steamship/client/__init__.py
@@ -1,3 +1,3 @@
-from .client import Steamship
+from .steamship import Steamship
 
 __all__ = ["Steamship"]
