+++ src/steamship/data/plugin/__init__.py
@@ -1,0 +1,18 @@
+from .hosting import HostingCpu, HostingEnvironment, HostingMemory, HostingTimeout, HostingType
+from .plugin import Plugin, PluginAdapterType, PluginTargetType, PluginType
+from .plugin_instance import PluginInstance
+from .plugin_version import PluginVersion
+
+__all__ = [
+    "Plugin",
+    "PluginVersion",
+    "PluginInstance",
+    "HostingMemory",
+    "HostingTimeout",
+    "HostingCpu",
+    "HostingEnvironment",
+    "HostingType",
+    "PluginType",
+    "PluginAdapterType",
+    "PluginTargetType",
+]
